[feat - #63] JDBC batch repository 생성
